WIP refaktorere litt innlogget skjema
